put pr to master

+++ src/main/java/com/qiniu/storage/persistent/FileRecorder.java
@@ -1,6 +1,7 @@
 package com.qiniu.storage.persistent;
 
 import com.qiniu.storage.Recorder;
+import com.qiniu.util.StringUtils;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -14,6 +15,7 @@ import java.util.Date;
  */
 public final class FileRecorder implements Recorder {
     private final File directory;
+    private static final String SPLIT = "_.-^ \b";
 
     /**
      * 断点记录文件保存的目录
@@ -52,12 +54,16 @@ public final class FileRecorder implements Recorder {
      * @param data 上传文件的进度数据
      */
     @Override
-    public void set(String key, byte[] data) {
-        File f = new File(directory, hash(key));
+    public synchronized void set(String key, byte[] data) {
+        if (StringUtils.isNullOrEmpty(key)) {
+            return;
+        }
+        File f = new File(directory, key);
         FileOutputStream fo = null;
         try {
             fo = new FileOutputStream(f);
             fo.write(data);
+            fo.flush();
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -76,8 +82,11 @@ public final class FileRecorder implements Recorder {
      * @param key 上传文件进度文件保存名
      */
     @Override
-    public byte[] get(String key) {
-        File f = new File(directory, hash(key));
+    public synchronized byte[] get(String key) {
+        if (StringUtils.isNullOrEmpty(key)) {
+            return null;
+        }
+        File f = new File(directory, key);
         if (!f.exists()) {
             return null;
         }
@@ -109,7 +118,7 @@ public final class FileRecorder implements Recorder {
     }
 
     private boolean outOfDate(File f) {
-        return f.lastModified() + 1000 * 3600 * 24 * 2 < new Date().getTime();
+        return f.lastModified() + 1000 * 3600 * 24 * 5 < new Date().getTime();
     }
 
     /**
@@ -118,21 +127,30 @@ public final class FileRecorder implements Recorder {
      * @param key 上传文件进度文件保存名
      */
     @Override
-    public void del(String key) {
-        File f = new File(directory, hash(key));
+    public synchronized void del(String key) {
+        if (StringUtils.isNullOrEmpty(key)) {
+            return;
+        }
+        File f = new File(directory, key);
         f.delete();
     }
 
     @Override
     public String recorderKeyGenerate(String key, File file) {
-        return key + "_._" + file.getAbsolutePath();
+        return hash(key + SPLIT + file.getAbsolutePath() + SPLIT + file.lastModified());
+    }
+
+
+    @Override
+    public String recorderKeyGenerate(String bucket, String key, String contentDataSUID, String uploaderSUID) {
+        return hash(bucket + SPLIT + key + SPLIT + contentDataSUID + SPLIT + uploaderSUID);
     }
 
     private static String hash(String base) {
         try {
             MessageDigest digest = MessageDigest.getInstance("SHA-1");
             byte[] hash = digest.digest(base.getBytes());
-            StringBuffer hexString = new StringBuffer();
+            StringBuilder hexString = new StringBuilder();
 
             for (int i = 0; i < hash.length; i++) {
                 hexString.append(Integer.toString((hash[i] & 0xff) + 0x100, 16).substring(1));
@@ -141,6 +159,6 @@ public final class FileRecorder implements Recorder {
         } catch (Exception ex) {
             ex.printStackTrace();
         }
-        return null;
+        return "";
     }
 }
